# Optimize to accommodate multiple error msg