Test result:
(01/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.exceed_slot: STARTED
 (01/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.exceed_slot: PASS (31.94 s)
 (02/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.max_addr: STARTED
 (02/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.max_addr: PASS (39.96 s)
 (03/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.unexisted_node: STARTED
 (03/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.unexisted_node: PASS (40.11 s)
 (04/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.unexisted_path: STARTED
 (04/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.unexisted_path: PASS (39.74 s)
 (05/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.invalid_alignsize: STARTED
 (05/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.invalid_alignsize: PASS (40.24 s)
 (06/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.invalid_addr_type: STARTED
 (06/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.invalid_addr_type: PASS (39.99 s)
 (07/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.small_label: STARTED
 (07/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.small_label: PASS (40.00 s)
 (08/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.bigger_label: STARTED
 (08/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.bigger_label: PASS (39.29 s)
 (09/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.bigger_target_memory: STARTED
 (09/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.bigger_target_memory: PASS (39.56 s)
 (10/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.with_discard: STARTED
 (10/10) type_specific.io-github-autotest-libvirt.memory.devices.invalid_nvdimm.with_numa.with_discard: PASS (39.08 s)